backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.2
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
pillow==10.4.0
aiofiles==24.1.0
python-dotenv==1.0.1
pdfplumber==0.10.3
python-docx==1.1.0
oss2>=2.18.0
